Fix: Mark episode as watched feature and audit test coverage

+++ app/src/main/java/com/theupnextapp/domain/TraktShowRating.kt
@@ -22,7 +22,7 @@
 package com.theupnextapp.domain
 
 data class TraktShowRating(
-    val rating: Int?,
+    val rating: Double?,
     val votes: Int?,
     val distribution: HashMap<String, Int>?,
 )
